fix syntax highlighting for json code block
sphinx tries its best to highlight code blocks, but sometimes it just can't because it cannot tolerate any mistakes in the syntax. Making the code block an explicit `{code-block}` class and adding `:force: true` to the directives tells sphinx to highlight the best it can because the world's not perfect.