Use uv to create and manage hubDocs development environments
This changeset replaces the current setup and preview instructions with
uv-based instructions. The goal is to make it easier for contributors
to work locally--uv will transparently handle Python installs and virtual
environments.

While it would be possible to reduce the setup steps even further by adopting
uv's lockfile, I chose to retain a requirements.txt-based setup because
the output is human-readable (and because others on the team use different
tools, and all tools "speak" requirements.txt)