Update files flagged by the codespell pre-commit hook
This changeset contains spelling errors detected and corrected by
the codespell pre-commit hook. Keeping these in a separate commit gives
us the flexibility to pick and choose whether or not we want to commit changes
made by the other hooks (there are many: those will be in the next commit).