Update files modified by the trailing-whitespace pre-commit hook
This is a one-time update for existing hubDocs files that fail the "no
trailing spaces" pre-commit check. I recognize that this might be controversial
for markdown files, but the only reason we'd really need the two extra spaces
would be to start a new line without starting a new paragraph.

Removing the trailing whitespace means less noise when editing (depending on
your editor and how it's configured). That said, if this is a bridge too far,
we can yank this commit.