Bump all Python requirements
This replaces PR #275, which failed a readthedocs build due to
incompatible dependencies. We've seen this pattern in a few other Python-
based projects after enabling Dependabot. Because it's limited to 5 PRs
at a time, it sometimes submits updates to one package that cannot be
resolved in isolation from other package updates.

This changeset bumps all remaining Python dependencies at once. If this
pattern persists, we can tweak the Dependabot settings and use "group".