Remove note about custom names for the model-output directory (#294)
* Add mailto: prefix to hubverse mailing list

This resolves the following Sphinx build warning:
/Users/rsweger/code/hubDocs/docs/source/overview/support-consulting.md:28:
WARNING: 'myst' cross-reference target not found:
'hubverse+help@groups.io' [myst.xref_missing]

* Remove note about flexible name for the model-output directory

Although the corresponding schema change is not yet in place, we've decided
to standardize on model-output for the model output data (similar to how
we've standardized on target-data).

This changeset pre-emptively removes the note about flexible names so that
new users are encouraged to use model-output.